+++ src/main/java/vn/com/gsoft/products/model/dto/PhieuKiemKesReq.java
@@ -10,11 +10,11 @@ public class PhieuKiemKesReq extends BaseRequest {
     private Integer phieuNhapMaPhieuNhap;
     private Integer phieuXuatMaPhieuXuat;
     private String nhaThuocMaNhaThuoc;
-    private Integer userProfileUserId;
+    private Long userProfileUserId;
     private Boolean daCanKho;
     private Boolean active;
     private Integer soPhieu;
-    private Integer archivedId;
-    private Integer storeId;
+    private Long archivedId;
+    private Long storeId;
     private Date archivedDate;
 }
