+++ src/main/java/vn/com/gsoft/products/model/dto/WarehouseLocationReq.java
@@ -15,8 +15,8 @@ public class WarehouseLocationReq extends BaseRequest {
     private Date created;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd/MM/yyyy HH:mm:ss")
     private Date modified;
-    private Integer createBy;
-    private Integer modifieBy;
+    private Long createBy;
+    private Long modifieBy;
     private String descriptions;
 }
